requirements.txt
# add your package for development
django
qiniu
django-qiniu-storage
django-account-helper
